Rename data_factory to base_data
