+++ candystore/__init__.py
@@ -1,3 +1,3 @@
 """Factories for randomised AFL data sets for testing purposes."""
 
-from .data_factory import CandyStore
+from .base_data import CandyStore
