+++ candystore/base_data.py
